Trying to update autoconf tool chain
Also use defined autoconf variables correctly.

+++ src/common/common.h
@@ -1,6 +1,5 @@
 #ifndef COMMON_H
 #define COMMON_H
-#include <string.h>
 
 #include "structures.h"
 
@@ -8,7 +7,7 @@
    header file. This should make the other libraries independant of
    the hazard header file.
 */
-typedef long double xtended;
+typedef double xtended;
 
 struct params{
   double dnmrlo,dnmrhi;
@@ -23,10 +22,10 @@ struct hzropt {
   double ostep;
   double ordmax;
   double *orders,*swordr;
-  long *rstvec;
+  int *rstvec;
   struct rsttbl *rsttbl;
-  long fdigit,N,itrlmt,prtcod,iret,itrcnt,trmcod,consec;
-  long orderv,nosw,restct,bdlvar,bdlcnt;
+  int fdigit,N,itrlmt,prtcod,iret,itrcnt,trmcod,consec;
+  int orderv,nosw,restct,bdlvar,bdlcnt;
 };
 
 struct hzrstr {
@@ -37,40 +36,40 @@ struct hzrstr {
   double factor;
   double *Mu[4];
   double slike;
-  long mode,fixmu,ivar,nfncts;
-  long ierr;
+  int mode,fixmu,ivar,nfncts;
+  int ierr;
   logical weibul,truhes,newton,anhess,ihpass,angrad;
   logical fixg1,fixg3;
-  long nocor,nocov;
-  long *index;
+  int nocor,nocov;
+  int *index;
   double *lng1,*lng2,*lng3;
   double *lnsg1,*lnsg3;
   double *lng1ct,*lng2ct,*lng3ct;
   double *lng1st,*lng2st,*lng3st;
   double *lnmu1,*lnmu2,*lnmu3;
-  long *indxx,*indxp,*indxt;
+  int *indxx,*indxp,*indxt;
   double *d2llad,*d2ll,*d1ll1,*d1ll2,*d1ll3;
   double *d2ll11,*d2ll12,*d2ll13,*d2ll22,*d2ll23,*d2ll33;
   double *beta,*x;
-  long *indx,*indx2;
+  int *indx,*indx2;
   double *cov,*corr,*sd,*znorm,*pnorm,*wk;
-  long *flag;
+  int *flag;
   double *lvec,*x0,*xc,*xplus,*gc,*gplus,*svec,*yvec,*sx,*typx;
-  long pj1,pp1,pp7,mp;
+  int pj1,pp1,pp7,mp;
   char chgflg[12];
   double sle,sls;
-  long sw,bw,f,mov,nps,npq,done,only1,step,retry,lstvar,N;
+  int sw,bw,f,mov,nps,npq,done,only1,step,retry,lstvar,N;
   charVarName vstep,fstep;
   double zstep,pstep;
-  long swadel,hpsn,fdone;
-  long mxstop,mxstep,mxvars;
+  int swadel,hpsn,fdone;
+  int mxstop,mxstep,mxvars;
   double *stdErr,*zvalue,*pvalue,*qx,*qtols;
-  long *nx,*nm1dx,*moves,*mxmove,*flags;
-  long maxn,phase1,phase2,phase3;
+  int *nx,*nm1dx,*moves,*mxmove,*flags;
+  int maxn,phase1,phase2,phase3;
   charVarName *names,*namex;
   charLabel *labl,*lablx;
   logical label;
-  long cnt,cntp[4];
+  int cnt,cntp[4];
   struct hzropt opt;
 };
 
@@ -89,7 +88,7 @@ EXTERN struct namestr *T_ns;
 EXTERN struct params Params;
 EXTERN struct {
   xtended lnlike;
-  long I,Im1p7;
+  int I,Im1p7;
   double lmu[4];
   double d1llp[4],d2llij[4][4];
   enum {SETCOE,CONSTP,CONSRV,DLLIKE,D2LCON} entry;
