+++ src/common/hzf_col.h
@@ -1,6 +1,6 @@
 #ifndef HZF_COL_H
 #define HZF_COL_H
 
-void hzf_col(long column);
+void hzf_col(int column);
 
 #endif /* HZF_COL_H */
