+++ src/common/hzf_memget.h
@@ -1,7 +1,7 @@
 #ifndef HZF_MEMGET_H
 #define HZF_MEMGET_H
 
-void *hzf_memget(long leng);
+void *hzf_memget(int leng);
 
 #endif /* HZF_MEMGET_H */
