+++ src/common/hzf_put.h
@@ -1,6 +1,6 @@
 #ifndef HZF_PUT_H
 #define HZF_PUT_H
 
-void hzf_put(long leng,char *strg);
+void hzf_put(int leng,char *strg);
 
 #endif /* HZF_PUT_H */
