+++ src/common/hzflg1.h
@@ -1,6 +1,6 @@
 #ifndef HZFLG1_H
 #define HZFLG1_H
 
-void hzflg1(char *strg,long leng);
+void hzflg1(char *strg,int leng);
 
 #endif /* HZFLG1_H */
