+++ src/common/hzflg2.h
@@ -1,6 +1,6 @@
 #ifndef HZFLG2_H
 #define HZFLG2_H
 
-void hzflg2(char *strg1,long leng1,char *strg2,long leng2);
+void hzflg2(char *strg1,int leng1,char *strg2,int leng2);
 
 #endif /* HZFLG2_H */
