+++ src/common/hzfskp.h
@@ -1,6 +1,6 @@
 #ifndef HZFSKP_H
 #define HZFSKP_H
 
-void hzfskp(long count);
+void hzfskp(int count);
 
 #endif /* HZFSKP_H */
