+++ src/common/hzfxpc.h
@@ -1,6 +1,6 @@
 #ifndef HZFXPC_H
 #define HZFXPC_H
 
-void hzfxpc(char *strg,long leng,long column);
+void hzfxpc(char *strg,int leng,int column);
 
 #endif /* HZFXPC_H */
