+++ src/common/hzfxpf.h
@@ -1,6 +1,6 @@
 #ifndef HZFXPF_H
 #define HZFXPF_H
 
-void hzfxpf(double value,long places,long precsn,long column);
+void hzfxpf(double value,int places,int precsn,int column);
 
 #endif /* HZFXPF_H */
