+++ src/common/hzfxpi.h
@@ -1,6 +1,6 @@
 #ifndef HZFXPI_H
 #define HZFXPI_H
 
-void hzfxpi(long value,long places,long column);
+void hzfxpi(int value,int places,int column);
 
 #endif /* HZFXPI_H */
