+++ src/common/hzfxpr.h
@@ -1,6 +1,6 @@
 #ifndef HZFXPR_H
 #define HZFXPR_H
 
-void hzfxpr(char byte,long times,long column);
+void hzfxpr(char byte,int times,int column);
 
 #endif /* HZFXPR_H */
