+++ src/common/namfix.h
@@ -1,6 +1,6 @@
 #ifndef NAMFIX_H
 #define NAMFIX_H
 
-void namfix(char *varname,long varidx);
+void namfix(char *varname,int varidx);
 
 #endif /* NAMFIX_H */
