+++ src/common/notspec.h
@@ -2,5 +2,5 @@
 #define NOTSPEC_H
 #include "structures.h" /* Defines logical */
 
-logical notspec(long fldno);
+logical notspec(int fldno);
 #endif /* NOTSPEC_H */
