+++ src/common/stmtfld.h
@@ -1,6 +1,6 @@
 #ifndef STMTFLD_H
 #define STMTFLD_H
 
-double stmtfld(long fldno);
+double stmtfld(int fldno);
 
 #endif /* STMTFLD_H */
