+++ src/common/xexit.h
@@ -1,6 +1,6 @@
 #ifndef XEXIT_H
 #define XEXIT_H
 
-void xexit(long rc);
+void xexit(int rc);
 
 #endif /* XEXIT_H */
