+++ src/hazard/alocmem.c
@@ -1,15 +1,14 @@
+#include <string.h>
 #include <hzdinc.h>
 #include <hzf_memget.h>
 #include <hzf_log1.h>
 #include <hzfxit.h>
-#include <string.h>
-
 #include "hazard.h"
 
 /****************************************************************/
 /* ALOCMEM called from main                                     */
 void alocmem(void){
-  long p_cnt,count,j; 
+  int p_cnt,count,j; 
 
   /* Allocate analysis memory */
   p_cnt = 11+3*C->p;
@@ -22,10 +21,10 @@ void alocmem(void){
   conmin = hzf_memget(C->p*sizeof(double));
   conmax = hzf_memget(C->p*sizeof(double));
   comean = hzf_memget(C->p*sizeof(double));
-  conmis = hzf_memget(C->p*sizeof(long));
-  iearly = hzf_memget(C->p*sizeof(long));
-  iconst = hzf_memget(C->p*sizeof(long));
-  ilate = hzf_memget(C->p*sizeof(long));
+  conmis = hzf_memget(C->p*sizeof(int));
+  iearly = hzf_memget(C->p*sizeof(int));
+  iconst = hzf_memget(C->p*sizeof(int));
+  ilate = hzf_memget(C->p*sizeof(int));
   H->names = hzf_memget(p_cnt*VAR_NAME_LENGTH);
   H->namex = hzf_memget(p_cnt*VAR_NAME_LENGTH);
   H->labl = hzf_memget(p_cnt*LABEL_LENGTH);
@@ -38,17 +37,17 @@ void alocmem(void){
   H->qx = hzf_memget(p_cnt*sizeof(double));
   H->qtols = hzf_memget(p_cnt*sizeof(double));
   C->status = hzf_memget((p_cnt+1)*sizeof(short int));
-  pres6 = hzf_memget(p_cnt*sizeof(long));
-  H->index = hzf_memget(p_cnt*sizeof(long));
-  H->nx = hzf_memget(p_cnt*sizeof(long));
-  H->nm1dx = hzf_memget(p_cnt*sizeof(long));
-  H->moves = hzf_memget(p_cnt*sizeof(long));
-  H->mxmove = hzf_memget(p_cnt*sizeof(long));
-  H->flags = hzf_memget(p_cnt*sizeof(long));
+  pres6 = hzf_memget(p_cnt*sizeof(int));
+  H->index = hzf_memget(p_cnt*sizeof(int));
+  H->nx = hzf_memget(p_cnt*sizeof(int));
+  H->nm1dx = hzf_memget(p_cnt*sizeof(int));
+  H->moves = hzf_memget(p_cnt*sizeof(int));
+  H->mxmove = hzf_memget(p_cnt*sizeof(int));
+  H->flags = hzf_memget(p_cnt*sizeof(int));
   C->obs = hzf_memget((7+C->p)*C->Nobs*sizeof(double));
   names_ns = hzf_memget(p_cnt*sizeof(struct namestr *));
   O->orders = hzf_memget(p_cnt*sizeof(double));
-  O->rstvec = hzf_memget(p_cnt*sizeof(long));
+  O->rstvec = hzf_memget(p_cnt*sizeof(int));
   if(hzf_mem_need>0) {
     sprintf(msgbfr,"ERROR: Insufficient memory for Stage II; "
 	    "need %ldK additional storage",hzf_mem_need/1024);
