+++ src/hazard/cput.c
@@ -1,10 +1,11 @@
+#include <string.h>
 #include <hzfxpc.h>
 #include <hzfxpf.h>
 #include <hzfskp.h>
 
 #include "hazard.h"
 
-void cput(char *phasnm,long j){
+void cput(char *phasnm,int j){
   hzfxpc(phasnm,9,31);
   hzfxpc(risk[j],8,-1);
   hzfxpf(comean[j],12,99,-6);
