+++ src/hazard/cput_d.c
@@ -1,10 +1,11 @@
+#include <string.h>
 #include <hzfxpc.h>
 #include <hzfxpi.h>
 #include <hzfxpf.h>
 #include <hzfskp.h>
 #include "hazard.h"
 
-void cput_d(char *phasnm,long j){
+void cput_d(char *phasnm,int j){
   hzfxpc(phasnm,9,26);
   hzfxpc(risk[j],8,-1);
   hzfxpi(conmis[j],5,-6);
