+++ src/hazard/cput_d.h
@@ -1,6 +1,6 @@
 #ifndef CPUT_D_H
 #define CPUT_D_H
 
-void cput_d(char *phasnm,long j);
+void cput_d(char *phasnm,int j);
 #endif /* CPUT_D_H */
