+++ src/hazard/cput_dl.h
@@ -1,4 +1,4 @@
 #ifndef CPUT_DL_H
 #define CPUT_DL_H
-void cput_dl(char *phasnm,long j);
+void cput_dl(char *phasnm,int j);
 #endif /* CPUT_DL_H */
