+++ src/hazard/cput_l.c
@@ -1,9 +1,10 @@
+#include <string.h>
 #include "hazard.h"
 #include <hzfxpc.h>
 #include <hzfxpf.h>
 #include <hzfskp.h>
 
-void cput_l(char *phasnm,long j){
+void cput_l(char *phasnm,int j){
   hzfxpc(phasnm,9,10);
   hzfxpc(risk[j],8,-1);
   hzfxpc(H->labl[j+9],40,-6);
