+++ src/hazard/cput_l.h
@@ -1,5 +1,5 @@
 #ifndef CPUT_L_H
 #define CPUT_L_H
 
-void cput_l(char *phasnm,long j);
+void cput_l(char *phasnm,int j);
 #endif /* CPUT_L_H */
