+++ src/hazard/cput_v.h
@@ -1,6 +1,6 @@
 #ifndef CPUT_V_H
 #define CPUT_V_H
 
-void cput_v(long dashct,long dashat,void (*cputfn)(char *,long));
+void cput_v(int dashct,int dashat,void (*cputfn)(char *,int));
 #endif /* CPUT_V_H */
