+++ src/hazard/cput_vv.h
@@ -1,5 +1,5 @@
 #ifndef CPUT_VV_H
 #define CPUT_VV_H
 
-void cput_vv(char *phasnm,long *iary,void (*cputfn)(char *,long));
+void cput_vv(char *phasnm,int *iary,void (*cputfn)(char *,int));
 #endif /* CPUT_VV_H */
