+++ src/hazard/memclc2.c
@@ -1,8 +1,9 @@
+#include <string.h>
 #include "hazard.h"
 
 /****************************************************************/
 /* MEMCLC2 called from ALOCWRK                                  */
-void memclc2(long phasno,long xt1st,long *xt2nd){
+void memclc2(int phasno,int xt1st,long *xt2nd){
   if(C->phase[phasno]==1)
     *xt2nd = xt1st+C->Nobs;
   else
