+++ src/hazard/memclc2.h
@@ -1,4 +1,4 @@
 #ifndef MEMCLC2_H
 #define MEMCLC2_H
-void memclc2(long phasno,long xt1st,long *xt2nd);
+void memclc2(int phasno,int xt1st,int *xt2nd);
 #endif /* MEMCLC2_H */
