+++ src/hazard/outparm.h
@@ -1,6 +1,6 @@
 #ifndef OUTPARM_H
 #define OUTPARM_H
 
-void outparm(char *phase,long chgno,long parmno,double spcval,
+void outparm(char *phase,int chgno,int parmno,double spcval,
 	     double useval,char *estnam);
 #endif /* OUTPARM_H */
