+++ src/hazard/phasidx.h
@@ -1,4 +1,4 @@
 #ifndef PHASIDX_H
 #define PHASIDX_H
-void phasidx(long *i,long phasno,long spt,long ept);
+void phasidx(int *i,int phasno,int spt,int ept);
 #endif /* PHASIDX_H */
